Finish V1

+++ Designs_Patterns/Model/IngerdientsProps.cs
@@ -1,15 +1,23 @@
 ﻿using System;
 namespace Designs_Patterns.Model
 {
+	/// <summary>
+	/// class Ingredients 
+	/// </summary>
 	public class IngerdientsProps
     {
-		public IngerdientsProps(int quantity, FoodProps foodProps)
+		/// <summary>
+		/// Constructeur 
+		/// </summary>
+		/// <param name="quantity"></param>
+		/// <param name="foodProps"></param>
+		public IngerdientsProps(double quantity, FoodProps foodProps)
 		{
 			this.quantity = quantity;
 			this.foodProps = foodProps;
 		}
 
-		public int quantity { get; set; }
+		public double quantity { get; set; }
 
 		public FoodProps foodProps { get; set; }
 	}
