+++ Designs_Patterns/Model/PizzaProps.cs
@@ -5,7 +5,7 @@ namespace Designs_Patterns
 {
 	public class PizzaProps
 	{
-		public PizzaProps(string title, List<IngerdientsProps> recipe, int prix)
+		public PizzaProps(string title, List<IngerdientsProps> recipe, double prix)
 		{
 			this.title = title;
 			this.recipe = recipe;
@@ -16,7 +16,7 @@ namespace Designs_Patterns
 
 		public string title { get; set; }
 
-		public int prix { get; set; }
-	}
+		public double prix { get; set; }
+    }
 }
